requirements.txt
opencv_contrib_python==4.5.3.56
scipy==1.7.1
pandas==1.3.1
torchvision==0.9.0
tqdm==4.62.2
numpy==1.21.2
torch==1.8.0
matplotlib==3.4.2
Pillow==8.4.0
